Allow to define logging format and class via env variables
It is hard to override default logging output format without building new image.
In theory, it can be done by passing format string as system property to `JAVA_OPTS` variable,
but since the format often contains whitespaces and $ sign and `JAVA_OPTS` is passed through `eval`
and at then end it is expanded in `exec` without double quotes, it is extremely hard to make it
passing properly.