Update logging configuration after changes in distribution
Set default value for LOGGING_LEVEL in log4j2 properties.

In https://github.com/hazelcast/hazelcast/pull/20528 we remove setting
the default value in the scripts and moved it do the log4j2 config file.

In Docker image we use specific config file without rolling file
appender (only log to console). The changes in the main distribution
need to be applied to this config files as well.