Create GH release in job without build matrix [HZ-943]
When we release docker images the build uses a build matrix, which runs
2 similar jobs - one for full image the other for slim image. Both try
to create the GH release and the second action fails because the release
has already been created, causing the whole workflow run to fail.

This commit moves the release creation (and update of readme, which
needs to be done only once) to a separate job in the workflow, without
the build matrix.