Allow `simple-smoke-test.sh` to be run multiple times  (#745)
`.github/scripts/simple-smoke-test.sh` creates and runs a container with
a default name (e.g. `hazelcast-oss`), but doesn't remove the container
afterwards, because later we query Docker for the logs.

This means when executed locally, subsequent invocations fail as the
container already exists.

Changed to remove the container if it already exists, before creation.