Use .github/release_type for controlling releases of the OSS/EE [5.4.z] (#752)
Backport of https://github.com/hazelcast/hazelcast-docker/pull/750

In order to control which editions (OSS/EE/ALL) should be released
`.github/release_type` file is introduced.

This file is required now for newly created git tags. It should contains
only one of the following values:
- OSS
- EE
- ALL

When a new tag is pushed we read the file to decide which docker image
editions should be released.
For manual runs of the `tag_image_push.yml` we ignore this file and use
`RELEASE_TYPE` input from the workflow

`.github/release_type` should be added during the release by the
automation. It's forbidden to add it master or maintenance branches
through PRs.