`xmllint` missing from `build-pr-custom-jdk` PR builder job [5.3.z]
`xmllint` is required since
https://github.com/hazelcast/hazelcast-docker/pull/747, but [a PR
builder execution
failed](https://github.com/hazelcast/hazelcast-docker/actions/runs/9698651682/job/26765696410)
because it wasn't available.

This is because while it's in the workflow, it's not in the specific
`build-pr-custom-jdk` _job_.

I'm unsure why this failure is intermittent as it should be continual.

(cherry picked from commit 3f9928489f5b43cb46026fc6a19f98ee9d8bb669)