Reduce duplication in scripts between test and released images - `HAZELCAST_ZIP_URL` [DI-156] [5.3.z]
The generation of the `HAZELCAST_ZIP_URL` parameter should be saved to
ensure consistency between images used for testing, and the image that's
ultimately deployed to ensure the test is as representative as possible.

In addition, a test has been added to check the type of distribution in
the image - EE or OSS - as a sanity check.

Fixes: [DI-156](https://hazelcast.atlassian.net/browse/DI-156)

[DI-156]:
https://hazelcast.atlassian.net/browse/DI-156?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

(cherry picked from commit 5581d1d5b3db4e75a85d322f6cc2bc478e862b2f)