When `RHEL` publish fails, log reason (`test_results`) (#845)
While investigating an [action
failure](https://github.com/hazelcast/hazelcast-docker/actions/runs/12407221911/attempts/1)
of an `RHEL` publish failure, the logs indicated that the image had
never published, but not _why_.

The publish response includes a link to an `test_results` result which
can [provide more
information](https://catalog.redhat.com/api/containers/docs/endpoints/RESTGetTestResultsByImageId.html).

Updated the script to log these results in the case of failure.

I've gone backwards and forwards on the implementation - _ideally_ the
publish _request_ should be polled rather than the the presence of an
image, but would require significant refactoring of the script which is
difficult to test due to its destructive nature.
In fact, in isolation this change is pointless - we build from tags so
the usage of this change will only gradually increase. But it helps to
document _how_ we get investigate subsequent failures. Ideally once next
investigated, the root cause can be identified and resolved.