Investigate Docker RHEL 5.5.2 build issue [DI-401] (#868)
**Details**
This fixes RHEL publishing timeout issues raised with RedHat support
[ticket](https://connect.redhat.com/support/partner-acceleration-desk/#/case/04042093).
RH support confirmed the scan can take anything from a few mins to 3h.
To account for this and to make script more robust, this PR introduces
the following changes:

1. `delete_unpublished_images()` - deletes all unpublished images for a
given version before scanning is requested. After deletion, the script
verifies the images are truly deleted, otherwise fails. We currently
have to do this manually from RedHat console so this automates the task
2. Updated `TIMEOUT_IN_MINS` in YML to 240 which equates to 4hrs

Fixes: [DI-401](https://hazelcast.atlassian.net/browse/DI-401)

---------

Co-authored-by: Jack Green <jack.green@hazelcast.com>
Co-authored-by: Łukasz Dziedziul <lukasz.dziedziul@hazelcast.com>