Ensure RHEL tests use a valid version of Management Center [DI-422]
The RHEL publish tests the build by connecting an Management Center instance to it.

In https://github.com/hazelcast/hazelcast-docker/pull/847 this was changed to use a _matching_ version of Hazelcast & Management Center - which isn't correct in all cases (e.g. `5.5.2` Hazelcast & `5.6.0` Management Center)

Before this change, it used the latest version of Management Center - derived by inspecting the tags of the (archived) https://github.com/hazelcast/management-center-openshift repo - but the last tag here was `5.3.2` so it wasn't really the latest at all.

In [consultation with the Management Center team](https://hazelcast.slack.com/archives/C9S7BV74Z/p1738850975557879), latest is the best option.

Changes:
- choses latest based on tags
- tags are scraped from the `hazelcast/management-center` repo
- tags are queried using the `gh` cli tool to avoid checking out the whole repo

I don't think we should backport this change, as while it's not correct in earlier versions, it doesn't cause an issues.

Fixes: [DI-422](https://hazelcast.atlassian.net/browse/DI-422)