Fix `tag_image_push` for `push` triggers
The issue described in https://github.com/hazelcast/hazelcast-docker/pull/846 _also_ affects `tag_image_push`.

Specifically, when run on `push`, `RELEASE_VERSION` is empty - because it gets it value set to `${{ env.HZ_VERSION }}` which is _pre-parsed_ as an empty value and not updated when we update `HZ_VERSION` earlier.

Post merge actions:
- [ ] backport