Use `README` from `master` when updating Docker Hub description (#918)
As discussed in
https://github.com/hazelcast/hazelcast-docker/pull/915#issuecomment-2720752695,
when a tag is pushed, we update the Docker Hub description based on the
`README`... in that tag. Which means when a patch for an older version
is released, we effectively roll back the description to an older
version.

Changes:
- on tag `push`, update the Docker Hub description in a separate job
based on `master`
- bonus - because it was in a matrix we _were_ doing it multiple times
- when the `README` or associated architecture is updated, rebuild as
well
- move manual `README` updates from `tag_image_push` to the new action,
which simplifies a *lot*
- convert Docker readme update to matrix to avoid duplicating action
specification

Tested
[here](https://github.com/hazelcast/hazelcast-docker/actions/runs/13855972414).