Separate RedHat build and test to increase build reliability [DI-438] (#938)
We have recurrent build failures caused by flakey RedHat + Docker +
Openshift tests - and due to some quirks, failure to rebuild an image
causes the _existing_ image (not just the replacement) to be removed.

The value this test offers is low - the Docker image is already being
tested (the RedHat & EE images are the same) elsewhere, and testing the
Helm chart + OpenShift is out of scope for this repository.

This PR removes this test as part of the build, but is still available
to be run manually.

[Example
execution](https://github.com/hazelcast/hazelcast-docker/actions/runs/14641182669/job/41083775318).

Post-merge items:
- [ ] backport (will be painful)
- [ ] retag

Fixes: [DI-438](https://hazelcast.atlassian.net/browse/DI-438)

[DI-438]:
https://hazelcast.atlassian.net/browse/DI-438?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ