Automate retagging process when release branches change [5.5.5] (#960)
Backport of https://github.com/hazelcast/hazelcast-docker/pull/955

When a release branch (`5.5.5`) is changed, the corresponding tag needs
to be recreated also - which is a manual step.

Instead, we can automate this process via a GitHub action that:
- identifies if the `push` is to a release branch (not a maintenance
branch)
- checks a tag _already_ exists (the scope of this action is to retag,
not to _release_ a branch)
- perform the existing retagging procedure

Note - this is a sticking-plaster over an imperfect process.

Testing performed in a scratch-repo:
- skipping:
-
[`main`](https://github.com/JackPGreen/backport-test/actions/runs/14894149859)
- [maintenance
branch](https://github.com/JackPGreen/backport-test/actions/runs/14894189254)
- [retagging on `5.5.6` release
branch](https://github.com/JackPGreen/backport-test/actions/runs/14894111197/job/41832985389)