Check RedHat service status on `push` failure (#970)
The `push` [was
failing](https://hazelcast.slack.com/archives/C06EQG13Z7H/p1747148998201369)
but it took a while for us to realise the problem was
[upstream](https://status.redhat.com/incidents/k7kvfvgfrbdf):
> Quay.io has been moved to read-only mode while we work to implement a
fix. During this time Pulls will continue to work, however, Pushes will
be disabled until a fix is implemented. There is currently no timeline
for Push restoration

We can use the [RedHat status API](https://status.redhat.com/api) to
query the status and give a heads-up if this is likely to be related.

[Example
output](https://github.com/hazelcast/hazelcast-docker/actions/runs/15001432702):

![image](https://github.com/user-attachments/assets/9d8f3ac0-faba-447f-a430-0de5b5959de2)


Post-merge actions:
- [ ] backport

---------

Co-authored-by: Nishaat Rajabali <12186256+nishaatr@users.noreply.github.com>