Fix RedHat `Image you are trying to publish did not pass the certification test` error [DI-469] (#968)
Since removing the smoke test in
https://github.com/hazelcast/hazelcast-docker/pull/938, the build has
become less stable due to exposing unreliable behaviour.

Specifically, when querying the state of the image in RedHat, we
_assume_ that as we've deleted `unpublished` images, when we query for
(say) `5.5.1-jdk17`, that _must_ be the one were trying to publish.

However, the "unpublishing" doesn't work as expected, which means the
image we get when querying isn't what we expect and the rest of the
process goes awry.
More specifically, an _existing_ published image (`5.5.1-jdk17`) has
data _like_:
```
{
  "total": 1,
  "data": [
    {
      "repositories": [
        {
          "published": false,
          "push_date": "2025-05-10T10:39:48+00:00"
          }
        },
        {
          "published": true,
          "published_date": "2025-05-10T10:44:16.083000+00:00",
          "push_date": "2025-05-10T10:44:16.083000+00:00"
        }
      ]
    }
  ]
}
```

That's one _image_ with _both_ published and not published. So when we
request to delete any unpublished images via the API, were going to
delete any image that _contains_ an unpublished component - which _may_
help explain why _published_ images are going missing.

Instead, we should _explicitly_ provide the ID of the image we want to
query the state for - and this sidesteps the requirement to delete
unpublished images as part of the build altogether.

Changes:
- stop removing unpublished images before doing anything
- extract the `IMAGE_ID` from the preflight output
- use `IMAGE_ID` over `VERSION` (effectively the tag name)
- remove sort of `get_image` result
   - should only be one result
- _assuming_ the first item in a list is the one you want is always
going to be racey regardless of how you sort it

[Example (debug)
execution](https://github.com/hazelcast/hazelcast-docker/actions/runs/14944560112/job/41986227335).

Closes: https://github.com/hazelcast/hazelcast-docker/pull/967 (would
cause too many conflicts)
Fixes: [DI-469](https://hazelcast.atlassian.net/browse/DI-469)

Post-merge items:
- [ ] backport

[DI-469]:
https://hazelcast.atlassian.net/browse/DI-469?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ