Prevent action-failure Slack notifications for non-automated actions [DI-464] (#984)
When actions fail, they send a Slack notification - but this happens
even for actions triggered via PR, which makes them noisy.

Updated to notify only for automated actions, where there is no real
initiator - as otherwise the GitHub `actor` has _already_ been notified
via email by GitHub.

For actions that run on a schedule or are triggered from other
workflows, this change is trivial - but for those that are _always_
triggered "manually" via `workflow_dispatch`, we can only check the
`actor` to reason if it's an automated action.

Removed notification entirely from `rhel-smoke-test` as this is _only
ever_ triggered manually.

Fixes: [DI-464](https://hazelcast.atlassian.net/browse/DI-464)

[DI-464]:
https://hazelcast.atlassian.net/browse/DI-464?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ