Rename `HZ_SNAPSHOT_INTERNAL_DOCKER_REGISTRY` secret (#993)
The name of the secret is confusing - the secret actually provides the
URL of the server in which the repository is contained, not the name of
the repository.

Name chosen as we authenticate with it using `JFROG`-labelled
credentials:
https://github.com/hazelcast/hazelcast-docker/blob/66909c17bff790e8eb9dc639693c9488728b3f42/.github/workflows/oss_latest_snapshot_push.yml#L101-L106

Raised from [PR
feedback](https://github.com/hazelcast/hazelcast-docker/pull/991#discussion_r2140621328).