Fix `version.functions_tests.sh` (#1007)
The assertion says that `xmllint` was removed in `5.4.2`, but [it
wasn't](https://github.com/hazelcast/hazelcast-docker/blob/v5.4.2/.github/actions/install-xmllint/action.yml)
so it fails.

I'm not sure what the assertion is testing, but we already have coverage
for the positive case of `get_last_version_with_file` so I think it's ok
to remove.

Raised from
https://github.com/hazelcast/hazelcast-docker/pull/1006#issuecomment-3024007418