Centralise Docker setup actions (qemu + buildx)
The `docker/setup-buildx-action@v3` with `version: v0.5.1` is duplicated in 7 actions. The version used for the build system should be specified only once - extracted.

Note - this version is [old](https://github.com/docker/buildx/releases/tag/v0.5.1) which is what prompted centralising to make it easier to explore upgrading in future.