Centralise `variants` matrix
The `variants` matrix (describing available distribution `classifiers` - either `slim` or regular) is duplicated in 4 places and can be centralised - and at the same time can centralise the computation of the `suffix` derived from the `variant`.

Also removed the `include` matrix which _appears_ to duplicate `variant` but is unused.

This action _could_ have been externalised to `docker-actions` but seemed overengineered, instead access directly on `master` - but until merged, the builds will fail.