Execute `tag_image_push` actions from `master` [DI-581] (#1039)
The `push` actions are automatically triggered on `push` events to their
respective branches/tags, but the actions that are triggered are run on
the branch themselves - which means that to support `N` maintenance
versions, we end up supporting `N` copies of the build scripts.

This also exposes further issues when trying to rebuild the images on a
scheduled basis - GitHub does not support dynamic usage of re-usable
workflows, so instead we have to trigger the actions on the specified
branches via `gh` asynchronously and then lose all visibility of the
execution outcome:

https://github.com/hazelcast/hazelcast-docker/blob/36410467219c8215d538d4f28b4d0e2ed7a3cdb7/.github/workflows/check-base-images.yml#L96-L97

Update to execute from `master`, including using the latest scripts,
actions etc - but to checkout the specified `ref` to use for the
underlying `Dockerfile`.

Changes:
- `push` trigger removed
   - will require upstream pipeline changes
   -  actions *must* be manually triggered
- `SOURCE_REF` parameter added
   -  specify the target `hazelcast-docker` branch
- `HZ_VERSION` input removed
    - computed dynamically based on the specified branch
- optional `RELEASE_VERSION` input removed
- allowed releasing `hazelcast-docker` with a separate lifecycle (e.g.
`5.5.1.1`), outside of the release of `hazelcast` (e.g. at `5.5.1`)
- unsure if ever been used in production, adds cognitive overhead to
pipeline

[Example
execution](https://github.com/hazelcast/hazelcast-docker/actions/runs/16818172462/).

Fixes: [DI-581](https://hazelcast.atlassian.net/browse/DI-581)

[DI-581]:
https://hazelcast.atlassian.net/browse/DI-581?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ