Make RedHat publish serial [DI-570]
The RedHat publish is intermittently failing, most recently [here](https://github.com/hazelcast/hazelcast-docker/actions/runs/16285479358).

Specifically, the publish as asynchronous - we submit a publish request and poll for publish completion, which _intermittently_ never happens.

Further investigation [in conjunction with RedHat support](https://connect.redhat.com/support/partner-acceleration-desk/#/case/04139855) identified that the reason they aren't getting published is that _something_ is subsequently requesting the image to be _un_published. This is _probably_ a `sync-tags` request which has the (server-side) side-effect of unpublishing any subsequently-orphaned images - i.e. if an image has had it's tags removed and is no longer "accessible", it is removed.

However, although we can hypothesise _how_ they are removed, we do not understand _why_ - the images do not overlap and it _appears_ to be a race condition on the RedHat backend. This issue is more prevalent with scheduled image rebuilds where we build many versions and images in parallel.

As such, we will no longer publish RedHat images in parallel.

We have multiple layers of concurrency:
- per-JDK
   - we build JDK variants (`17`, `21` etc) of the _same_ version in parallel in the matrix
   - `max-parallel` set to restrict this
- per-version
   - multiple versions (`5.4`, `5.5`) are triggered in parallel by the scheduled image rebuild, or manually
   - `concurrency` group set to restrict this - [note that GitHub supports at most two jobs within the same concurrency group](https://docs.github.com/en/actions/how-tos/writing-workflows/choosing-what-your-workflow-does/control-the-concurrency-of-workflows-and-jobs#using-concurrency-in-different-scenarios) (one running, one pending) and any additional jobs will be rejected - not queued as you might expect. This is acceptable as we _currently_ only have two versions _in_ RedHat, but _if_ that increased we may need to consider an alternative solution, additional restrictions etc.

Fixes: [DI-570](https://hazelcast.atlassian.net/browse/DI-570)