Avoid needlessly overwriting `HZ_VERSION`
The `Dockerfile`s have an `HZ_VERSION` arg.

We read these using [docker-actions/get-hz-versions](https://github.com/hazelcast/docker-actions/tree/master/get-hz-versions) to determine the Hazelcast version for a given `Dockerfile`.

...but then when we _build_ the `Dockerfile`, we override the `HZ_VERSION` argument with the value _we just read_ from the `Dockerfile`.

Updated to stop overriding this property.

The _only_ case when we might want to override it is in the PR builder where we build an OS image from `master`, but use a _published_ (e.g. `5.5`) version of Hazelcast... but the OS `Dockerfile` doesn't actually _use_ the `HZ_VERSION` for anything anyway so it's pointless.