Merge EE & OS build jobs (#1061)
`tag_image_push` builds EE and OS images - but the structure is
_something like_:
- build OS test image
- test OS test image
- build real OS image
- build EE test image
- test EE test image
- build EE real image

The commands to build and test _should_ be basically identical, and
reducing duplication avoids them become out-of-sync.

Refactored to run (basically) the same commands to perform the
build-test/test/build-real operations with the EE-or-OS flag coming from
a matrix.

The net result of this is -30LOC and the PR builder drops from
[7m43s](https://github.com/hazelcast/hazelcast-docker/actions/runs/17179499550)
->
[5m47s](https://github.com/hazelcast/hazelcast-docker/actions/runs/17236039849)
(1⁄3 faster) because the OS/EE builds can be done in parallel.

Pre-merge actions:
- [ ] trigger an _actual_ push