Refactor `publish-rhel.sh` (#1083)
### Simplify
The script calls functions in order, most steps only proceed if a
certain condition is met. The subsequent step then immediately checks
the same condition before proceeding which is redundant.

This is an optimization, it's not a problem.

Example:
`wait_for_container_scan`:

https://github.com/hazelcast/hazelcast-docker/blob/6729095dbdf6d8ba17f5eaf9b6d5dfbf6f321b99/.github/scripts/publish-rhel.sh#L68-L69
Immediately followed by `publish_the_image`:

https://github.com/hazelcast/hazelcast-docker/blob/6729095dbdf6d8ba17f5eaf9b6d5dfbf6f321b99/.github/scripts/publish-rhel.sh#L103-L106


`wait_for_container_publish`:

https://github.com/hazelcast/hazelcast-docker/blob/6729095dbdf6d8ba17f5eaf9b6d5dfbf6f321b99/.github/scripts/publish-rhel.sh#L138-L146
Immediately followed by `sync_tags`:

https://github.com/hazelcast/hazelcast-docker/blob/6729095dbdf6d8ba17f5eaf9b6d5dfbf6f321b99/.github/scripts/publish-rhel.sh#L183-L189

Finally merged `publish_the_image` & `sync_tags` functions as _now_ they
are the same.

### `get_image`
`get_image` took in filter parameters (e.g. `deleted`, `published`),
because _prior_ to
https://github.com/hazelcast/hazelcast-docker/pull/968, it was filtering
by image _name_ and multiple results were expected.

Now that we use `IMAGE_ID`, there should only be one response.

This also means rather than filtering `published` on the server-side, we
can get the image and check the `published` status of the response -
which allows us to log it for debug purposes.

### `wait_for_container_publish`

In the event of a timeout waiting for a publish, we log the test results
since https://github.com/hazelcast/hazelcast-docker/pull/845
We should _also_ add information on the status of the (publish) request,
as this is what helped us identify the root cause of the RedHat publish
issues

---------

Co-authored-by: Nishaat Rajabali <12186256+nishaatr@users.noreply.github.com>