Fix RedHat publish by adding dummy unique tag to images [DI-605] (#1088)
[RedHat builds regularly
fail](https://github.com/hazelcast/hazelcast-docker/actions/runs/17416749534).

After further investigation [with their support
team](https://connect.redhat.com/support/partner-acceleration-desk/#/case/04237767),
the root cause is a bug on the RedHat side that prevents the publish.

Specifically, after a publish, it incorrectly identifies an image as
having _no_ tags and then immediately removes it. This _appears_ to be
some kind of race condition triggered when a new image replaces a
previous images tags, and has no unique tags of it's own, but
investigation on their side is ongoing.

Their proposed workaround is to add a permanent, _unique_ tag, and that
another user has the same problem and uses this solution. I have
implemented this by augmenting the tags with an _additional_ tag - the
digest of the image:

<img width="1943" height="722" alt="image"
src="https://github.com/user-attachments/assets/15272032-d271-4c2a-a3a5-03ef9390b18e"
/>


[Example
execution](https://github.com/hazelcast/hazelcast-docker/actions/runs/17502970827).

Note for reviewers - this change requires / builds on top of
https://github.com/hazelcast/hazelcast-docker/pull/1082 - we need to get
the digest of the image _before_ we can push so the steps have to be
seperate.

Fixes: [DI-605](https://hazelcast.atlassian.net/browse/DI-605)

[DI-605]:
https://hazelcast.atlassian.net/browse/DI-605?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ