Fix `check-base-images` `outdated` check (#1098)
When checking if an image is outdated, we check if the output of the
action is _present_, not what it actually _is_. Which means we always
end up rebuilding.

Specifically, in [this
build](https://github.com/hazelcast/hazelcast-docker/actions/runs/17892193720/job/50874374488),
`outdated` is `false` (because it _just_ rebuilt the images) but it
rebuilds again anyway.

[Example execution from this branch showing it _correctly_ skipping the
rebuild](https://github.com/hazelcast/hazelcast-docker/actions/runs/17895389872).

Also fixed the "if no rebuilds required skip the rebuilds" check [which
was setting the job status to failure in this
case](https://github.com/hazelcast/hazelcast-docker/actions/runs/17895353812).