Make PR Builder use images from branch (#1104)
In https://github.com/hazelcast/hazelcast-docker/pull/1039 the PR
builder was changed from:
- building an image from the PR branches' `Dockerfile`
- using the Hazelcast binary of a released version

To:
- building an image from a released versions' `Dockerfile`
- using the Hazelcast binary of a released version

Which means that Dockerfile changes could not be tested.

The idea of using a released version (over a `SNAPSHOT`) was to allow
execution without access to credentials guarding private repositories,
but this _already_ doesn't work as other parts of the PR builder need
access to enterprise license keys to test the enterprise image.