Download distribution ZIPs using Maven (#1111)
Currently we calculate the URL of the ZIP in the Maven repository and
pass that into the image when it's built.

It would be simpler if we could instead ask Maven to download instead:
- we don't need to pass (potentially sensitive) URLs around
- we don't need to construct URLs by string-concatenation
- instead can pass individual components (e.g. `classifier`) to Maven
and it can figure it out
   - we also don't need to _test_ that stuff
- makes it easier to change the repository dynamically (e.g. to an
`environment`) as it doesn't need to be passed-around

This was the intention when the scripts were written:
https://github.com/hazelcast/hazelcast-docker/blob/c5664c013f43433a156a200f1398ff83e817c4e7/.github/scripts/ee-build.functions.sh#L3-L4

But _at the time_ these were being called _inside_ the Docker image
while being built - now we call the scripts _outside_ (in the action)
there is no such restriction.