Use autopublish for RedHat (#1125)
We _currently_:
1. push images to the RedHat scan registry
2. scan with `preflight` - getting the images `ID`
3. send a publish request for that `ID`
4. poll-and-wait for that `ID` to be published
5. perform a `sync-tags` operation

While investigating
https://github.com/hazelcast/management-center-docker/pull/230 I
realised that the MC RedHat image was being released at (2) and the
subsequent steps were redundant, because the MC RedHat component has
`autopublish` enabled.

RedHat support define `autopublish` as:
> Autopublish operation will sync the catalog data with your images
submitted to the public registry and ensure that the image with the
right tag is published on the catalog.
{...}
Additionally, enabling auto-publish could save you some time.

As such we've already _been_ testing `autopublish` with MC for a long
time.

We can also skip the `sync-tags` operation, defined by RedHat support
as:
> Sync-tags operation ensures that the new image that is pushed is
consistent across all registries. For example, if you are copying from
Docker to Quay, the sync-tag operation will ensure that the image tag is
consistent in both registries.
{...}
Sync-tags operation will synchronize tags in both registries. This means
that if you are copying from Docker to Quay, the images will have the
same tags in both registries. Since Docker is "parent" (the one that is
copied from), the image in Quay will have the same tag as in the Docker
registry.

This _doesn't appear_ to apply in our case - although we _are_ copying
from Docker to Quay, we are going _via_ the scan registry and manually
copying every (applicable) tag.

Finally, an additional problem with our _current_ approach is that we
send a publish request for a _single_ `ID` - but in the case of MC with
multi-architecture image, each image has it's _own unique_ `ID` - this
wasn't visible in `hazelcast-docker` as the images are
single-architecture.

Changes:
- remove anything beyond `preflight` (2)

Tested in sandbox:
<img width="1867" height="480" alt="image"
src="https://github.com/user-attachments/assets/e8ca3c69-0bfe-4e58-872d-2f4d0899fdc7"
/>