Update `Building Your Hazelcast Image` documentation [DI-664] (#1147)
The docs explain:

https://github.com/hazelcast/hazelcast-docker/blob/2ac469bc4e2756ea35c2c746fea543f3557f90aa/README.md?plain=1#L277-L290
Which was added in
https://github.com/hazelcast/hazelcast-docker/pull/111.

Since then this has been superceeded - in
https://github.com/hazelcast/hazelcast-docker/commit/a1cd3eb12f94a98a78a0dcc3882e7885b7ea0f9d,
`jar`s were replaced with distribution `zip`s and in
https://github.com/hazelcast/hazelcast-docker/pull/365 support was added
to use a local distribution `zip` as a mechanism for deploying a custom
Hazelcast binary.

The instructions don't work - if you include a `jar`, it's copied to
`HZ_HOME` (_not_ `lib` or `bin`) and it has no effect.

As such, removed.

Note - this documentation covers replacing the _Hazelcast_ binary only -
_additional_ content (e.g. adding a _new_ `jar`) to the image is
[documented
separately](https://github.com/hazelcast/hazelcast-docker#extending-hazelcast-base-image).