Remove `.github/release_type` files (#1160)
The files were added to allow executions without context (i.e. on tag
push) where EE/OS/ALL inputs could not be supplied.

However, not only is this no longer the case, but they override the
_actual_ inputs supplied.

Specifically, when rebuilding the Docker images on schedule, we trigger
`tag_image_push` with an input of `EE`. However, internally it then
overrides this with the content of `release_type` - for current releases
that's _also_ been `EE` (e.g. `5.4.3`, `5.5.8` etc)... but for `5.6.0`
the content of that file is `ALL` so the scheduled rebuilt ended up
(re)building OSS too.

Fixed by removing this (now redundant) functionality.