Fix RedHat publish (#1163)
In
https://github.com/hazelcast/hazelcast-docker/commit/b2c7ed7a71c86c6f76f22f2bbda88feb32944afa,
the restriction on if should be published to RedHat was refined to check
the condition was actually `true`.

...however, `resolve-editions` actually outputted:
https://github.com/hazelcast/docker-actions/blob/d225708d0c8b05c0c15c356b1455b5fc59f58536/resolve-editions/build.functions.sh#L16-L20

And therefore RedHat publish was skipped.

While we _could_ fix the check, this behaviour is very confusing and
slipped through a code review unnoticed. Instead, we should use
conventional `true`/`false` output constants.