Fix scheduled rebuild (#1161)
The change in https://github.com/hazelcast/hazelcast-docker/pull/1157
did not consider scheduled events, where the new `FORCE_REBUILD` input
will be `null`.

Also while debugging I noticed that the downloaded artefact used as part
of the resolution of the versions to rebuild was not read correctly when
there was a single value, meaning _nothing_ was triggered. This is
related to https://github.com/actions/download-artifact?v5---whats-new.
Updated config to make behaviour consistent, and updated implementation
to suit.

Example executions:
-
[single](https://github.com/hazelcast/hazelcast-docker/actions/runs/19372569057/job/55431917979)
-
[multiple](https://github.com/hazelcast/hazelcast-docker/actions/runs/19373359979)