Migrate `check-if-latest-lts-release` to `docker-actions`, remove reliance on `.github/lts` file [DI-687][DI-580] (#1178)
This change is comprised of three parts:

### Bug & fix
The _existing_ `check-if-latest-lts-release` implementation relied on
the existence of a `.github/lts` file to signify an LTS release, and
absence the inverse.
In https://github.com/hazelcast/hazelcast-docker/pull/1039, the default
checked-out branch was changed from the branch to be built to `master` -
but `check-if-latest-lts-release` was left alone, looking in the default
location. This meant it _never_ found a `.github/lts` file, assuming
non-LTS, [even in the case of a real LTS
version](https://github.com/hazelcast/hazelcast-docker/actions/runs/19818436231/job/56776593442).

Assuming the _lack_ of an LTS file as proof of non-LTS is error prone -
and the concept of an "LTS file" is a bit hacky anyway.

### Refactoring
Rather than pass through paths to allow the
`check-if-latest-lts-release` implementation to look in the _correct_
location for `.github/lts`, we can refactor to avoid this lookup
entirely.
Instead, we can query the LTS version using the `hazelcast` `pom.xml`:

https://github.com/hazelcast/hazelcast-mono/blob/30e3147adde0b3d42e6f4ac99e4c16bbc6caa51d/hazelcast/hazelcast-parent/pom.xml#L80

The `IS_LTS_OVERRIDE` input parameter has also been removed - this
_used_ to allow overriding the LTS release (regardless of whether it was
actually LTS or not) so long as version was _after_ the last LTS version
(e.g. making `5.6.0` LTS even though it's not, as long as it's after
`5.5.8` that is).
This complexity is difficult to re-implement, although a simplified
version (...simply override LTS _regardless_) could be added in future
if required. But based on LTS discussions, unlikely.

Notes:
- the tests have hardcoded LTS versions
   - this is brittle, but necessary
- i'm unsure of the ordering between `hazelcast` `pom.xml` being
updated, and the `hazelcast-docker` tags being created - if the latter
is before the former, LTS will not be correctly identified
   - scheduled rebuilds will address this
- it's only an issue for a _new_ LTS major version - this is unlikely to
be an issue going forwards

### Migration
To demonstrate the solution works, a test is required.
There's already a GitHub Actions assertion architecture in
`docker-actions`, hence it was easier to migrate the action to
`docker-actions` and use that architecture for new tests to demonstrate
that `is_latest_lts` is set appropriately.

[Example execution showing LTS correctly resolving for LTS (5.5) and
non-LTS (5.4/5.6)
versions](https://github.com/hazelcast/hazelcast-docker/actions/runs/19874689180).

Fixes: [DI-687](https://hazelcast.atlassian.net/browse/DI-687)
_Partially_ addresses:
[DI-580](https://hazelcast.atlassian.net/browse/DI-580)


[DI-687]:
https://hazelcast.atlassian.net/browse/DI-687?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[DI-580]:
https://hazelcast.atlassian.net/browse/DI-580?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[Example execution showing LTS correctly resolving for LTS (5.5) and
non-LTS (5.4/5.6)
versions](https://github.com/hazelcast/hazelcast-docker/actions/runs/19874651296).